# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "description": "Devuelve un JWT válido junto al resumen del usuario.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credenciales válidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSuccessDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas"
          },
          "403": {
            "description": "Cuenta deshabilitada (isActive = false)"
          }
        },
        "summary": "Inicio de sesión",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/accept-invitation": {
      "post": {
        "description": "Endpoint público. Crea el usuario con la contraseña elegida. No emite JWT; el cliente debe hacer login después.",
        "operationId": "AuthController_acceptInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta creada; el cliente debe ir a login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInvitationSuccessDto"
                }
              }
            }
          },
          "404": {
            "description": "Invitación no válida o ya utilizada"
          },
          "409": {
            "description": "Email ya registrado"
          }
        },
        "summary": "Aceptar invitación y crear cuenta",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "Requiere el encabezado Authorization: Bearer <token>.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usuario actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserSummaryDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Perfil del usuario autenticado",
        "tags": [
          "Autenticación"
        ]
      },
      "patch": {
        "description": "Actualmente permite establecer `mobilePhone`. Requiere Authorization: Bearer.",
        "operationId": "AuthController_patchMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserSummaryDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Actualizar perfil del usuario autenticado",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/me/password": {
      "patch": {
        "description": "En primer login (mustChangePassword) no se exige currentPassword. Tras el cambio, mustChangePassword pasa a false.",
        "operationId": "AuthController_patchPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserSummaryDto"
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o contraseña actual incorrecta"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cambiar contraseña del usuario autenticado",
        "tags": [
          "Autenticación"
        ]
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "description": "Envía un código OTP al correo si la cuenta existe y está activa. Siempre responde con el mismo mensaje genérico.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitud procesada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordSuccessDto"
                }
              }
            }
          }
        },
        "summary": "Solicitar código de recuperación de contraseña",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/verify-reset-code": {
      "post": {
        "description": "Valida el código sin consumirlo. El front puede usarlo para avanzar a la pantalla de nueva contraseña.",
        "operationId": "AuthController_verifyResetCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyResetCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Código válido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResetCodeSuccessDto"
                }
              }
            }
          }
        },
        "summary": "Verificar código OTP de recuperación",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "description": "Valida el código nuevamente, lo consume y establece la nueva contraseña.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordSuccessDto"
                }
              }
            }
          }
        },
        "summary": "Restablecer contraseña con código OTP",
        "tags": [
          "Autenticación"
        ]
      }
    },
...
    "/api/companies": {
      "get": {
        "description": "Solo admin de plataforma.",
        "operationId": "AdminCompaniesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol distinto de admin"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Listar todas las empresas",
        "tags": [
          "Empresas (Admin)"
        ]
      },
      "post": {
        "description": "Crea una ficha de empresa sin dueño de producto. Luego se invita a un business.",
        "operationId": "AdminCompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol distinto de admin"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Crear empresa",
        "tags": [
          "Empresas (Admin)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cotizador API",
    "description": "Contrato OpenAPI 3 para el front (`cotizador-app`).\n\nPrefijo global: `/api`. Auth: header `Authorization: Bearer <jwt>` (scheme `access-token`).\nJSON vivo: `GET /api/openapi.json`. UI: `/api/docs`.\nGuía de pantallas, roles y labels: `docs/contrato-front.md`.\n\nErrores: `{ statusCode, message, traceId? }`. `message` puede ser `string` o `string[]` (validación).\nMontos en CLP (número). Fechas de negocio `YYYY-MM-DD`; timestamps ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token obtenido en POST /api/auth/login"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@ejemplo.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Contraseña del usuario",
            "example": "********",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthUserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "mobilePhone": {
            "type": "string",
            "description": "Teléfono de contacto",
            "example": "+56912345678"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "business",
              "common"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "mobilePhone",
          "role",
          "isActive",
          "mustChangePassword",
          "createdAt",
          "updatedAt"
        ]
      },
      "LoginSuccessDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserSummaryDto"
          },
          "token": {
            "type": "string",
            "description": "JWT de acceso"
          },
          "expiresIn": {
            "type": "number",
            "description": "Segundos hasta la expiración del token",
            "example": 900
          }
        },
        "required": [
          "user",
          "token",
          "expiresIn"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token en claro recibido en el correo de invitación"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "MiClaveSegura1!"
          },
          "mobilePhone": {
            "type": "string",
            "example": "+56 9 1234 5678"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "AcceptInvitationSuccessDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Cuenta creada. Ya puedes iniciar sesión."
          }
        },
        "required": [
          "message"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "mobilePhone": {
            "type": "string",
            "description": "Teléfono de contacto; cadena vacía borra el valor",
            "example": "+56912345678"
          },
          "name": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Obligatoria si mustChangePassword es false; omitible en primer login"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "example": "NuevaClave123!"
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@ejemplo.cl"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordSuccessDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Si el correo existe en nuestro sistema, recibirás un código de verificación en breve."
          }
        },
        "required": [
          "message"
        ]
      },
      "VerifyResetCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@ejemplo.cl"
          },
          "code": {
            "type": "string",
            "description": "Código OTP de 6 dígitos enviado al correo",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "VerifyResetCodeSuccessDto": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "valid"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@ejemplo.cl"
          },
          "code": {
            "type": "string",
            "description": "Código OTP de 6 dígitos enviado al correo",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "example": "NuevaClave123!"
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
      "ResetPasswordSuccessDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Contraseña actualizada correctamente"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "CompanyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Mi Empresa SpA"
          },
          "rut": {
            "type": "string",
            "example": "76.123.456-7"
          },
          "address": {
            "type": "object",
            "example": "Av. Principal 123",
            "nullable": true
          },
          "city": {
            "type": "object",
            "example": "Santiago",
            "nullable": true
          },
          "contact": {
            "type": "object",
            "example": "+56 9 1234 5678",
            "nullable": true
          },
          "logoUrl": {
            "type": "object",
            "example": "https://cotizador-storage.sfo3.digitaloceanspaces.com/companies/uuid/logo.webp",
            "nullable": true,
            "description": "URL pública del logo (WebP en Spaces)"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "rut",
          "address",
          "city",
          "contact",
          "logoUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpsertCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mi Empresa SpA"
          },
          "rut": {
            "type": "string",
            "example": "76.123.456-7"
          },
          "address": {
            "type": "string",
            "example": "Av. Principal 123"
          },
          "city": {
            "type": "string",
            "example": "Santiago"
          },
          "contact": {
            "type": "string",
            "example": "+56 9 1234 5678"
          }
        },
        "required": [
          "name",
          "rut"
        ]
      },
...
    }
  }
}